Suggest using a high value for recurringExpiry
Since recurringExpiry is a required value and most recurring
transactions don't have a defined expiry date, suggest that the value is
sufficiently high.
It is my strong belief that no ACS/Issuer will disallow transactions
after the recurringExpiry of the authentication has expired.

I believe it was added by EMVCo because they saw it as another data point.